Started refactoring gaze estimation demo. Blocker: Script node bindings for get/set Sequence Num
